+++ src/PERI/fix_peri_neigh.cpp
@@ -397,7 +397,7 @@ void FixPeriNeigh::setup(int /*vflag*/)
 double FixPeriNeigh::memory_usage()
 {
   int nmax = atom->nmax;
-  int bytes = nmax * sizeof(int);
+  double bytes = nmax * sizeof(int);
   bytes += (double)nmax*maxpartner * sizeof(tagint);
   bytes += (double)nmax*maxpartner * sizeof(double);
   if (isVES) {
